requirements.txt
# AUTOBOT System Requirements

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Logging
python-json-logger>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
pandas-ta>=0.3.14b

# API Clients
python-binance>=1.0.19

# WebSocket (CRITICAL for real-time data)
websockets>=12.0

# Database/Cache
redis>=5.0.0

# Notifications
python-telegram-bot>=20.0

# Async
asyncio-mqtt>=0.13.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
